Crate spin_world

Re-exports

pub use fermyon::spin as v1;

pub use fermyon::spin2_0_0 as v2;

Modules

exports

fermyon

spin

wasi

wasi_otel

Structs

Host_

Auto-generated bindings for an instance a component which implements the world host.

Host_Indices

Auto-generated bindings for index of the exports of host.

Host_Pre

Auto-generated bindings for a pre-instantiated version of a component which implements the world host.

LinkOptions

Link-time configurations.

Constants

MAX_HOST_BUFFERED_BYTES

Maximum allowed size of a host-buffered database query result, HTTP request or response body, or similar.

Attribute Macros

async_trait