spin_world::fermyon::spin::key_value
Trait Host

pub trait Host: Send {
    // Required methods
    fn open<'life0, 'async_trait>(
        &'life0 mut self,
        name: String,
    ) -> Pin<Box<dyn Future<Output = Result<Result<Store, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get<'life0, 'async_trait>(
        &'life0 mut self,
        store: Store,
        key: String,
    ) -> Pin<Box<dyn Future<Output = Result<Result<Vec<u8>, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set<'life0, 'async_trait>(
        &'life0 mut self,
        store: Store,
        key: String,
        value: Vec<u8>,
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn delete<'life0, 'async_trait>(
        &'life0 mut self,
        store: Store,
        key: String,
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn exists<'life0, 'async_trait>(
        &'life0 mut self,
        store: Store,
        key: String,
    ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_keys<'life0, 'async_trait>(
        &'life0 mut self,
        store: Store,
    ) -> Pin<Box<dyn Future<Output = Result<Result<Vec<String>, Error>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn close<'life0, 'async_trait>(
        &'life0 mut self,
        store: Store,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn open<'life0, 'async_trait>( &'life0 mut self, name: String, ) -> Pin<Box<dyn Future<Output = Result<Result<Store, Error>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Open the store with the specified name.

If name is “default”, the default store is opened. Otherwise, name must refer to a store defined and configured in a runtime configuration file supplied with the application.

error::no-such-store will be raised if the name is not recognized.

fn get<'life0, 'async_trait>( &'life0 mut self, store: Store, key: String, ) -> Pin<Box<dyn Future<Output = Result<Result<Vec<u8>, Error>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the value associated with the specified key from the specified store.

error::invalid-store will be raised if store is not a valid handle to an open store, and error::no-such-key will be raised if there is no tuple for key in store.

fn set<'life0, 'async_trait>( &'life0 mut self, store: Store, key: String, value: Vec<u8>, ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Set the value associated with the specified key in the specified store, overwriting any existing value.

error::invalid-store will be raised if store is not a valid handle to an open store.

fn delete<'life0, 'async_trait>( &'life0 mut self, store: Store, key: String, ) -> Pin<Box<dyn Future<Output = Result<Result<(), Error>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Delete the tuple with the specified key from the specified store.

error::invalid-store will be raised if store is not a valid handle to an open store. No error is raised if a tuple did not previously exist for key.

fn exists<'life0, 'async_trait>( &'life0 mut self, store: Store, key: String, ) -> Pin<Box<dyn Future<Output = Result<Result<bool, Error>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return whether a tuple exists for the specified key in the specified store.

error::invalid-store will be raised if store is not a valid handle to an open store.

fn get_keys<'life0, 'async_trait>( &'life0 mut self, store: Store, ) -> Pin<Box<dyn Future<Output = Result<Result<Vec<String>, Error>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return a list of all the keys in the specified store.

error::invalid-store will be raised if store is not a valid handle to an open store.

fn close<'life0, 'async_trait>( &'life0 mut self, store: Store, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Close the specified store.

This has no effect if store is not a valid handle to an open store.