spin_world::fermyon::spin
Module sqlite

Structs

QueryResult

A result of a query

RowResult

A set of values for each of the columns in a query-result

Enums

Error

The set of errors which may be raised by functions in this interface

Value

Traits

GetHost

Host

Functions

add_to_linker

add_to_linker_get_host

Type Aliases

Connection

A handle to an open sqlite instance