spin_world::fermyon::spin::sqlite
Trait Host

pub trait Host: Send {
    // Required methods
    fn open<'life0, 'async_trait>(
        &'life0 mut self,
        database: String,
    ) -> Pin<Box<dyn Future<Output = Result<Connection, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn execute<'life0, 'async_trait>(
        &'life0 mut self,
        conn: Connection,
        statement: String,
        parameters: Vec<Value>,
    ) -> Pin<Box<dyn Future<Output = Result<QueryResult, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn close<'life0, 'async_trait>(
        &'life0 mut self,
        conn: Connection,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn convert_error(&mut self, err: Error) -> Result<Error>;
}

Required Methods

fn open<'life0, 'async_trait>( &'life0 mut self, database: String, ) -> Pin<Box<dyn Future<Output = Result<Connection, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Open a connection to a named database instance.

If database is “default”, the default instance is opened.

error::no-such-database will be raised if the name is not recognized.

fn execute<'life0, 'async_trait>( &'life0 mut self, conn: Connection, statement: String, parameters: Vec<Value>, ) -> Pin<Box<dyn Future<Output = Result<QueryResult, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Execute a statement returning back data if there is any

fn close<'life0, 'async_trait>( &'life0 mut self, conn: Connection, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Close the specified connection.

fn convert_error(&mut self, err: Error) -> Result<Error>