spin_world::spin::sqlite::sqlite

Trait HostConnection

pub trait HostConnection: Send {
    // Required methods
    fn open(
        &mut self,
        database: String,
    ) -> impl Future<Output = Result<Resource<Connection>, Error>> + Send;
    fn execute(
        &mut self,
        self_: Resource<Connection>,
        statement: String,
        parameters: Vec<Value>,
    ) -> impl Future<Output = Result<QueryResult, Error>> + Send;
    fn last_insert_rowid(
        &mut self,
        self_: Resource<Connection>,
    ) -> impl Future<Output = Result<i64>> + Send;
    fn changes(
        &mut self,
        self_: Resource<Connection>,
    ) -> impl Future<Output = Result<u64>> + Send;
    fn drop(
        &mut self,
        rep: Resource<Connection>,
    ) -> impl Future<Output = Result<()>> + Send;
}

Required Methods

fn open( &mut self, database: String, ) -> impl Future<Output = Result<Resource<Connection>, Error>> + Send

Open a connection to a named database instance.

If database is “default”, the default instance is opened.

error::no-such-database will be raised if the name is not recognized.

fn execute( &mut self, self_: Resource<Connection>, statement: String, parameters: Vec<Value>, ) -> impl Future<Output = Result<QueryResult, Error>> + Send

Execute a statement returning back data if there is any

fn last_insert_rowid( &mut self, self_: Resource<Connection>, ) -> impl Future<Output = Result<i64>> + Send

The SQLite rowid of the most recent successful INSERT on the connection, or 0 if there has not yet been an INSERT on the connection.

fn changes( &mut self, self_: Resource<Connection>, ) -> impl Future<Output = Result<u64>> + Send

The number of rows modified, inserted or deleted by the most recently completed INSERT, UPDATE or DELETE statement on the connection.

fn drop( &mut self, rep: Resource<Connection>, ) -> impl Future<Output = Result<()>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.