spin_world::wasi::cli0_3_0_rc_2025_09_16::stdout

Trait HostWithStore

pub trait HostWithStore: HasData + Send {
    // Required method
    fn write_via_stream<T>(
        accessor: &Accessor<T, Self>,
        data: StreamReader<u8>,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
}

Required Methods

fn write_via_stream<T>( accessor: &Accessor<T, Self>, data: StreamReader<u8>, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Write the given stream to stdout.

If the stream’s writable end is dropped this function will either return success once the entire contents of the stream have been written or an error-code representing a failure.

Otherwise if there is an error the readable end of the stream will be dropped and this function will return an error-code.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.