spin_world::wasi::clocks0_3_0_rc_2025_09_16::timezone

Trait Host

pub trait Host: Send {
    // Required methods
    fn display(
        &mut self,
        when: Datetime,
    ) -> impl Future<Output = Result<TimezoneDisplay>> + Send;
    fn utc_offset(
        &mut self,
        when: Datetime,
    ) -> impl Future<Output = Result<i32>> + Send;
}

Required Methods

fn display( &mut self, when: Datetime, ) -> impl Future<Output = Result<TimezoneDisplay>> + Send

Return information needed to display the given datetime. This includes the UTC offset, the time zone name, and a flag indicating whether daylight saving time is active.

If the timezone cannot be determined for the given datetime, return a timezone-display for UTC with a utc-offset of 0 and no daylight saving time.

fn utc_offset( &mut self, when: Datetime, ) -> impl Future<Output = Result<i32>> + Send

The same as display, but only return the UTC offset.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.