spin_world::wasi::filesystem0_3_0_rc_2025_09_16::types

Trait HostDescriptor

pub trait HostDescriptor: Send {
    // Required methods
    fn read_via_stream(
        &mut self,
        self_: Resource<Descriptor>,
        offset: Filesize,
    ) -> impl Future<Output = Result<(StreamReader<u8>, FutureReader<Result<(), ErrorCode>>)>> + Send;
    fn drop(
        &mut self,
        rep: Resource<Descriptor>,
    ) -> impl Future<Output = Result<()>> + Send;
}

Required Methods

fn read_via_stream( &mut self, self_: Resource<Descriptor>, offset: Filesize, ) -> impl Future<Output = Result<(StreamReader<u8>, FutureReader<Result<(), ErrorCode>>)>> + Send

Return a stream for reading from a file.

Multiple read, write, and append streams may be active on the same open file and they do not interfere with each other.

This function returns a stream which provides the data received from the file, and a future providing additional error information in case an error is encountered.

If no error is encountered, stream.read on the stream will return read-status::closed with no error-context and the future resolves to the value ok. If an error is encountered, stream.read on the stream returns read-status::closed with an error-context and the future resolves to err with an error-code.

Note: This is similar to pread in POSIX.

fn drop( &mut self, rep: Resource<Descriptor>, ) -> impl Future<Output = Result<()>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.