spin_world::wasi::filesystem0_3_0_rc_2025_09_16::types

Trait HostDescriptorWithStore

pub trait HostDescriptorWithStore: HasData + Send {

    // Required methods
    fn write_via_stream<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        data: StreamReader<u8>,
        offset: Filesize,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn append_via_stream<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        data: StreamReader<u8>,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn advise<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        offset: Filesize,
        length: Filesize,
        advice: Advice,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn sync_data<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn get_flags<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
    ) -> impl Future<Output = Result<Result<DescriptorFlags, ErrorCode>>> + Send;
    fn get_type<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
    ) -> impl Future<Output = Result<Result<DescriptorType, ErrorCode>>> + Send;
    fn set_size<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        size: Filesize,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn set_times<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        data_access_timestamp: NewTimestamp,
        data_modification_timestamp: NewTimestamp,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn read_directory<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
    ) -> impl Future<Output = Result<(StreamReader<DirectoryEntry>, FutureReader<Result<(), ErrorCode>>)>> + Send;
    fn sync<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn create_directory_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        path: String,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn stat<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
    ) -> impl Future<Output = Result<Result<DescriptorStat, ErrorCode>>> + Send;
    fn stat_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        path_flags: PathFlags,
        path: String,
    ) -> impl Future<Output = Result<Result<DescriptorStat, ErrorCode>>> + Send;
    fn set_times_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        path_flags: PathFlags,
        path: String,
        data_access_timestamp: NewTimestamp,
        data_modification_timestamp: NewTimestamp,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn link_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        old_path_flags: PathFlags,
        old_path: String,
        new_descriptor: Resource<Descriptor>,
        new_path: String,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn open_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        path_flags: PathFlags,
        path: String,
        open_flags: OpenFlags,
        flags: DescriptorFlags,
    ) -> impl Future<Output = Result<Result<Resource<Descriptor>, ErrorCode>>> + Send;
    fn readlink_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        path: String,
    ) -> impl Future<Output = Result<Result<String, ErrorCode>>> + Send;
    fn remove_directory_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        path: String,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn rename_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        old_path: String,
        new_descriptor: Resource<Descriptor>,
        new_path: String,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn symlink_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        old_path: String,
        new_path: String,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn unlink_file_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        path: String,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn is_same_object<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        other: Resource<Descriptor>,
    ) -> impl Future<Output = Result<bool>> + Send;
    fn metadata_hash<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
    ) -> impl Future<Output = Result<Result<MetadataHashValue, ErrorCode>>> + Send;
    fn metadata_hash_at<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<Descriptor>,
        path_flags: PathFlags,
        path: String,
    ) -> impl Future<Output = Result<Result<MetadataHashValue, ErrorCode>>> + Send;

}

Required Methods

fn write_via_stream<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, data: StreamReader<u8>, offset: Filesize, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Return a stream for writing to a file, if available.

May fail with an error-code describing why the file cannot be written.

It is valid to write past the end of a file; the file is extended to the extent of the write, with bytes between the previous end and the start of the write set to zero.

This function returns once either full contents of the stream are written or an error is encountered.

Note: This is similar to pwrite in POSIX.

fn append_via_stream<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, data: StreamReader<u8>, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Return a stream for appending to a file, if available.

May fail with an error-code describing why the file cannot be appended.

This function returns once either full contents of the stream are written or an error is encountered.

Note: This is similar to write with O_APPEND in POSIX.

fn advise<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, offset: Filesize, length: Filesize, advice: Advice, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Provide file advisory information on a descriptor.

This is similar to posix_fadvise in POSIX.

fn sync_data<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Synchronize the data of a file to disk.

This function succeeds with no effect if the file descriptor is not opened for writing.

Note: This is similar to fdatasync in POSIX.

fn get_flags<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, ) -> impl Future<Output = Result<Result<DescriptorFlags, ErrorCode>>> + Send

Get flags associated with a descriptor.

Note: This returns similar flags to fcntl(fd, F_GETFL) in POSIX.

Note: This returns the value that was the fs_flags value returned from fdstat_get in earlier versions of WASI.

fn get_type<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, ) -> impl Future<Output = Result<Result<DescriptorType, ErrorCode>>> + Send

Get the dynamic type of a descriptor.

Note: This returns the same value as the type field of the fd-stat returned by stat, stat-at and similar.

Note: This returns similar flags to the st_mode & S_IFMT value provided by fstat in POSIX.

Note: This returns the value that was the fs_filetype value returned from fdstat_get in earlier versions of WASI.

fn set_size<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, size: Filesize, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Adjust the size of an open file. If this increases the fileโs size, the extra bytes are filled with zeros.

Note: This was called fd_filestat_set_size in earlier versions of WASI.

fn set_times<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, data_access_timestamp: NewTimestamp, data_modification_timestamp: NewTimestamp, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Adjust the timestamps of an open file or directory.

Note: This is similar to futimens in POSIX.

Note: This was called fd_filestat_set_times in earlier versions of WASI.

fn read_directory<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, ) -> impl Future<Output = Result<(StreamReader<DirectoryEntry>, FutureReader<Result<(), ErrorCode>>)>> + Send

Read directory entries from a directory.

On filesystems where directories contain entries referring to themselves and their parents, often named . and .. respectively, these entries are omitted.

This always returns a new stream which starts at the beginning of the directory. Multiple streams may be active on the same directory, and they do not interfere with each other.

This function returns a future, which will resolve to an error code if reading full contents of the directory fails.

fn sync<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Synchronize the data and metadata of a file to disk.

This function succeeds with no effect if the file descriptor is not opened for writing.

Note: This is similar to fsync in POSIX.

fn create_directory_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, path: String, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Create a directory.

Note: This is similar to mkdirat in POSIX.

fn stat<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, ) -> impl Future<Output = Result<Result<DescriptorStat, ErrorCode>>> + Send

Return the attributes of an open file or directory.

Note: This is similar to fstat in POSIX, except that it does not return device and inode information. For testing whether two descriptors refer to the same underlying filesystem object, use is-same-object. To obtain additional data that can be used do determine whether a file has been modified, use metadata-hash.

Note: This was called fd_filestat_get in earlier versions of WASI.

fn stat_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, path_flags: PathFlags, path: String, ) -> impl Future<Output = Result<Result<DescriptorStat, ErrorCode>>> + Send

Return the attributes of a file or directory.

Note: This is similar to fstatat in POSIX, except that it does not return device and inode information. See the stat description for a discussion of alternatives.

Note: This was called path_filestat_get in earlier versions of WASI.

fn set_times_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, path_flags: PathFlags, path: String, data_access_timestamp: NewTimestamp, data_modification_timestamp: NewTimestamp, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Adjust the timestamps of a file or directory.

Note: This is similar to utimensat in POSIX.

Note: This was called path_filestat_set_times in earlier versions of WASI.

fn link_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, old_path_flags: PathFlags, old_path: String, new_descriptor: Resource<Descriptor>, new_path: String, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Create a hard link.

Fails with error-code::no-entry if the old path does not exist, with error-code::exist if the new path already exists, and error-code::not-permitted if the old path is not a file.

Note: This is similar to linkat in POSIX.

fn open_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, path_flags: PathFlags, path: String, open_flags: OpenFlags, flags: DescriptorFlags, ) -> impl Future<Output = Result<Result<Resource<Descriptor>, ErrorCode>>> + Send

Open a file or directory.

If flags contains descriptor-flags::mutate-directory, and the base descriptor doesnโt have descriptor-flags::mutate-directory set, open-at fails with error-code::read-only.

If flags contains write or mutate-directory, or open-flags contains truncate or create, and the base descriptor doesnโt have descriptor-flags::mutate-directory set, open-at fails with error-code::read-only.

Note: This is similar to openat in POSIX.

fn readlink_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, path: String, ) -> impl Future<Output = Result<Result<String, ErrorCode>>> + Send

Read the contents of a symbolic link.

If the contents contain an absolute or rooted path in the underlying filesystem, this function fails with error-code::not-permitted.

Note: This is similar to readlinkat in POSIX.

fn remove_directory_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, path: String, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Remove a directory.

Return error-code::not-empty if the directory is not empty.

Note: This is similar to unlinkat(fd, path, AT_REMOVEDIR) in POSIX.

fn rename_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, old_path: String, new_descriptor: Resource<Descriptor>, new_path: String, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Rename a filesystem object.

Note: This is similar to renameat in POSIX.

fn symlink_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, old_path: String, new_path: String, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Create a symbolic link (also known as a โsymlinkโ).

If old-path starts with /, the function fails with error-code::not-permitted.

Note: This is similar to symlinkat in POSIX.

fn unlink_file_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, path: String, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Unlink a filesystem object that is not a directory.

Return error-code::is-directory if the path refers to a directory. Note: This is similar to unlinkat(fd, path, 0) in POSIX.

fn is_same_object<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, other: Resource<Descriptor>, ) -> impl Future<Output = Result<bool>> + Send

Test whether two descriptors refer to the same filesystem object.

In POSIX, this corresponds to testing whether the two descriptors have the same device (st_dev) and inode (st_ino or d_ino) numbers. wasi-filesystem does not expose device and inode numbers, so this function may be used instead.

fn metadata_hash<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, ) -> impl Future<Output = Result<Result<MetadataHashValue, ErrorCode>>> + Send

Return a hash of the metadata associated with a filesystem object referred to by a descriptor.

This returns a hash of the last-modification timestamp and file size, and may also include the inode number, device number, birth timestamp, and other metadata fields that may change when the file is modified or replaced. It may also include a secret value chosen by the implementation and not otherwise exposed.

Implementations are encouraged to provide the following properties:

If the file is not modified or replaced, the computed hash value should usually not change.

If the object is modified or replaced, the computed hash value should usually change.

The inputs to the hash should not be easily computable from the computed hash.

However, none of these is required.

fn metadata_hash_at<T>( accessor: &Accessor<T, Self>, self_: Resource<Descriptor>, path_flags: PathFlags, path: String, ) -> impl Future<Output = Result<Result<MetadataHashValue, ErrorCode>>> + Send

Return a hash of the metadata associated with a filesystem object referred to by a directory descriptor and a relative path.

This performs the same hash computation as metadata-hash.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.