spin_world::wasi::http0_2_0
Module types

Structs

DnsErrorPayload

Defines the case payload type for DNS-error above:

FieldSizePayload

Defines the case payload type for HTTP-response-{header,trailer}-size above:

TlsAlertReceivedPayload

Defines the case payload type for TLS-alert-received above:

Enums

ErrorCode

These cases are inspired by the IANA HTTP Proxy Error Types: https://www.iana.org/assignments/http-proxy-status/http-proxy-status.xhtml#table-http-proxy-error-types

Fields

This following block defines the fields resource which corresponds to HTTP standard Fields. Fields are a common representation used for both Headers and Trailers.

FutureIncomingResponse

Represents a future which may eventaully return an incoming HTTP Response, or an error.

FutureTrailers

Represents a future which may eventaully return trailers, or an error.

HeaderError

This type enumerates the different kinds of errors that may occur when setting or appending to a fields resource.

IncomingBody

Represents an incoming HTTP Request or Response’s Body.

IncomingRequest

Represents an incoming HTTP Request.

IncomingResponse

Represents an incoming HTTP Response.

Method

This type corresponds to HTTP standard Methods.

OutgoingBody

Represents an outgoing HTTP Request or Response’s Body.

OutgoingRequest

Represents an outgoing HTTP Request.

OutgoingResponse

Represents an outgoing HTTP Response.

RequestOptions

Parameters for making an HTTP Request. Each of these parameters is currently an optional timeout applicable to the transport layer of the HTTP protocol.

ResponseOutparam

Represents the ability to send an HTTP Response.

Scheme

This type corresponds to HTTP standard Related Schemes.

Traits

GetHost

Host

HostFields

HostFutureIncomingResponse

HostFutureTrailers

HostIncomingBody

HostIncomingRequest

HostIncomingResponse

HostOutgoingBody

HostOutgoingRequest

HostOutgoingResponse

HostRequestOptions

HostResponseOutparam

Functions

add_to_linker

add_to_linker_get_host

Type Aliases

Duration

FieldKey

Field keys are always strings.

FieldValue

Field values should always be ASCII strings. However, in reality, HTTP implementations often have to interpret malformed values, so they are provided as a list of bytes.

Headers

Headers is an alias for Fields.

InputStream

IoError

OutputStream

Pollable

StatusCode

This type corresponds to the HTTP standard Status Code.

Trailers

Trailers is an alias for Fields.