spin_world::wasi::http0_2_0::types

Trait HostIncomingBody

pub trait HostIncomingBody: Send {
    // Required methods
    fn stream(
        &mut self,
        self_: Resource<IncomingBody>,
    ) -> impl Future<Output = Result<Result<Resource<InputStream>, ()>>> + Send;
    fn finish(
        &mut self,
        this: Resource<IncomingBody>,
    ) -> impl Future<Output = Result<Resource<FutureTrailers>>> + Send;
    fn drop(
        &mut self,
        rep: Resource<IncomingBody>,
    ) -> impl Future<Output = Result<()>> + Send;
}

Required Methods

fn stream( &mut self, self_: Resource<IncomingBody>, ) -> impl Future<Output = Result<Result<Resource<InputStream>, ()>>> + Send

Returns the contents of the body, as a stream of bytes.

Returns success on first call: the stream representing the contents can be retrieved at most once. Subsequent calls will return error.

The returned input-stream resource is a child: it must be dropped before the parent incoming-body is dropped, or consumed by incoming-body.finish.

This invariant ensures that the implementation can determine whether the user is consuming the contents of the body, waiting on the future-trailers to be ready, or neither. This allows for network backpressure is to be applied when the user is consuming the body, and for that backpressure to not inhibit delivery of the trailers if the user does not read the entire body.

fn finish( &mut self, this: Resource<IncomingBody>, ) -> impl Future<Output = Result<Resource<FutureTrailers>>> + Send

Takes ownership of incoming-body, and returns a future-trailers. This function will trap if the input-stream child is still alive.

fn drop( &mut self, rep: Resource<IncomingBody>, ) -> impl Future<Output = Result<()>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.