spin_world::wasi::http0_2_0::types
Trait HostIncomingRequest

pub trait HostIncomingRequest {
    // Required methods
    fn method<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<IncomingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Method>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn path_with_query<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<IncomingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn scheme<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<IncomingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Scheme>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn authority<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<IncomingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn headers<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<IncomingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Resource<Headers>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn consume<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<IncomingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Result<Resource<IncomingBody>, ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn drop<'life0, 'async_trait>(
        &'life0 mut self,
        rep: Resource<IncomingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn method<'life0, 'async_trait>( &'life0 mut self, self_: Resource<IncomingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Method>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the method of the incoming request.

fn path_with_query<'life0, 'async_trait>( &'life0 mut self, self_: Resource<IncomingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the path with query parameters from the request, as a string.

fn scheme<'life0, 'async_trait>( &'life0 mut self, self_: Resource<IncomingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Option<Scheme>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the protocol scheme from the request.

fn authority<'life0, 'async_trait>( &'life0 mut self, self_: Resource<IncomingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the authority from the request, if it was present.

fn headers<'life0, 'async_trait>( &'life0 mut self, self_: Resource<IncomingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Resource<Headers>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the headers associated with the request.

The returned headers resource is immutable: set, append, and delete operations will fail with header-error.immutable.

The headers returned are a child resource: it must be dropped before the parent incoming-request is dropped. Dropping this incoming-request before all children are dropped will trap.

fn consume<'life0, 'async_trait>( &'life0 mut self, self_: Resource<IncomingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Result<Resource<IncomingBody>, ()>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Gives the incoming-body associated with this request. Will only return success at most once, and subsequent calls will return error.

fn drop<'life0, 'async_trait>( &'life0 mut self, rep: Resource<IncomingRequest>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,