spin_world::wasi::http0_2_0::types
Trait HostOutgoingRequest

pub trait HostOutgoingRequest {
    // Required methods
    fn new<'life0, 'async_trait>(
        &'life0 mut self,
        headers: Resource<Headers>,
    ) -> Pin<Box<dyn Future<Output = Result<Resource<OutgoingRequest>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn body<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<OutgoingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Result<Resource<OutgoingBody>, ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn method<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<OutgoingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Method>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_method<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<OutgoingRequest>,
        method: Method,
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn path_with_query<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<OutgoingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_path_with_query<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<OutgoingRequest>,
        path_with_query: Option<String>,
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn scheme<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<OutgoingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Scheme>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_scheme<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<OutgoingRequest>,
        scheme: Option<Scheme>,
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn authority<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<OutgoingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_authority<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<OutgoingRequest>,
        authority: Option<String>,
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), ()>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn headers<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<OutgoingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Resource<Headers>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn drop<'life0, 'async_trait>(
        &'life0 mut self,
        rep: Resource<OutgoingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn new<'life0, 'async_trait>( &'life0 mut self, headers: Resource<Headers>, ) -> Pin<Box<dyn Future<Output = Result<Resource<OutgoingRequest>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Construct a new outgoing-request with a default method of GET, and none values for path-with-query, scheme, and authority.

headers is the HTTP Headers for the Request.

It is possible to construct, or manipulate with the accessor functions below, an outgoing-request with an invalid combination of scheme and authority, or headers which are not permitted to be sent. It is the obligation of the outgoing-handler.handle implementation to reject invalid constructions of outgoing-request.

fn body<'life0, 'async_trait>( &'life0 mut self, self_: Resource<OutgoingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Result<Resource<OutgoingBody>, ()>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the resource corresponding to the outgoing Body for this Request.

Returns success on the first call: the outgoing-body resource for this outgoing-request can be retrieved at most once. Subsequent calls will return error.

fn method<'life0, 'async_trait>( &'life0 mut self, self_: Resource<OutgoingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Method>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the Method for the Request.

fn set_method<'life0, 'async_trait>( &'life0 mut self, self_: Resource<OutgoingRequest>, method: Method, ) -> Pin<Box<dyn Future<Output = Result<Result<(), ()>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Set the Method for the Request. Fails if the string present in a method.other argument is not a syntactically valid method.

fn path_with_query<'life0, 'async_trait>( &'life0 mut self, self_: Resource<OutgoingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the combination of the HTTP Path and Query for the Request. When none, this represents an empty Path and empty Query.

fn set_path_with_query<'life0, 'async_trait>( &'life0 mut self, self_: Resource<OutgoingRequest>, path_with_query: Option<String>, ) -> Pin<Box<dyn Future<Output = Result<Result<(), ()>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Set the combination of the HTTP Path and Query for the Request. When none, this represents an empty Path and empty Query. Fails is the string given is not a syntactically valid path and query uri component.

fn scheme<'life0, 'async_trait>( &'life0 mut self, self_: Resource<OutgoingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Option<Scheme>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the HTTP Related Scheme for the Request. When none, the implementation may choose an appropriate default scheme.

fn set_scheme<'life0, 'async_trait>( &'life0 mut self, self_: Resource<OutgoingRequest>, scheme: Option<Scheme>, ) -> Pin<Box<dyn Future<Output = Result<Result<(), ()>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Set the HTTP Related Scheme for the Request. When none, the implementation may choose an appropriate default scheme. Fails if the string given is not a syntactically valid uri scheme.

fn authority<'life0, 'async_trait>( &'life0 mut self, self_: Resource<OutgoingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the HTTP Authority for the Request. A value of none may be used with Related Schemes which do not require an Authority. The HTTP and HTTPS schemes always require an authority.

fn set_authority<'life0, 'async_trait>( &'life0 mut self, self_: Resource<OutgoingRequest>, authority: Option<String>, ) -> Pin<Box<dyn Future<Output = Result<Result<(), ()>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Set the HTTP Authority for the Request. A value of none may be used with Related Schemes which do not require an Authority. The HTTP and HTTPS schemes always require an authority. Fails if the string given is not a syntactically valid uri authority.

fn headers<'life0, 'async_trait>( &'life0 mut self, self_: Resource<OutgoingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Resource<Headers>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the headers associated with the Request.

The returned headers resource is immutable: set, append, and delete operations will fail with header-error.immutable.

This headers resource is a child: it must be dropped before the parent outgoing-request is dropped, or its ownership is transfered to another component by e.g. outgoing-handler.handle.

fn drop<'life0, 'async_trait>( &'life0 mut self, rep: Resource<OutgoingRequest>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,