spin_world::wasi::http0_2_0::types

Trait HostOutgoingResponse

pub trait HostOutgoingResponse: Send {
    // Required methods
    fn new(
        &mut self,
        headers: Resource<Headers>,
    ) -> impl Future<Output = Result<Resource<OutgoingResponse>>> + Send;
    fn status_code(
        &mut self,
        self_: Resource<OutgoingResponse>,
    ) -> impl Future<Output = Result<StatusCode>> + Send;
    fn set_status_code(
        &mut self,
        self_: Resource<OutgoingResponse>,
        status_code: StatusCode,
    ) -> impl Future<Output = Result<Result<(), ()>>> + Send;
    fn headers(
        &mut self,
        self_: Resource<OutgoingResponse>,
    ) -> impl Future<Output = Result<Resource<Headers>>> + Send;
    fn body(
        &mut self,
        self_: Resource<OutgoingResponse>,
    ) -> impl Future<Output = Result<Result<Resource<OutgoingBody>, ()>>> + Send;
    fn drop(
        &mut self,
        rep: Resource<OutgoingResponse>,
    ) -> impl Future<Output = Result<()>> + Send;
}

Required Methods

fn new( &mut self, headers: Resource<Headers>, ) -> impl Future<Output = Result<Resource<OutgoingResponse>>> + Send

Construct an outgoing-response, with a default status-code of 200. If a different status-code is needed, it must be set via the set-status-code method.

headers is the HTTP Headers for the Response.

fn status_code( &mut self, self_: Resource<OutgoingResponse>, ) -> impl Future<Output = Result<StatusCode>> + Send

Get the HTTP Status Code for the Response.

fn set_status_code( &mut self, self_: Resource<OutgoingResponse>, status_code: StatusCode, ) -> impl Future<Output = Result<Result<(), ()>>> + Send

Set the HTTP Status Code for the Response. Fails if the status-code given is not a valid http status code.

fn headers( &mut self, self_: Resource<OutgoingResponse>, ) -> impl Future<Output = Result<Resource<Headers>>> + Send

Get the headers associated with the Request.

The returned headers resource is immutable: set, append, and delete operations will fail with header-error.immutable.

This headers resource is a child: it must be dropped before the parent outgoing-request is dropped, or its ownership is transfered to another component by e.g. outgoing-handler.handle.

fn body( &mut self, self_: Resource<OutgoingResponse>, ) -> impl Future<Output = Result<Result<Resource<OutgoingBody>, ()>>> + Send

Returns the resource corresponding to the outgoing Body for this Response.

Returns success on the first call: the outgoing-body resource for this outgoing-response can be retrieved at most once. Subsequent calls will return error.

fn drop( &mut self, rep: Resource<OutgoingResponse>, ) -> impl Future<Output = Result<()>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.