spin_world::wasi::http0_3_0_rc_2025_09_16

Module types

Structs

DnsErrorPayload

Defines the case payload type for DNS-error above:

FieldSizePayload

Defines the case payload type for HTTP-response-{header,trailer}-size above:

TlsAlertReceivedPayload

Defines the case payload type for TLS-alert-received above:

Enums

ErrorCode

These cases are inspired by the IANA HTTP Proxy Error Types: https://www.iana.org/assignments/http-proxy-status/http-proxy-status.xhtml#table-http-proxy-error-types

Fields

This following block defines the fields resource which corresponds to HTTP standard Fields. Fields are a common representation used for both Headers and Trailers.

HeaderError

This type enumerates the different kinds of errors that may occur when setting or appending to a fields resource.

Method

This type corresponds to HTTP standard Methods.

Request

Represents an HTTP Request.

RequestOptions

Parameters for making an HTTP Request. Each of these parameters is currently an optional timeout applicable to the transport layer of the HTTP protocol.

RequestOptionsError

This type enumerates the different kinds of errors that may occur when setting fields of a request-options resource.

Response

Represents an HTTP Response.

Scheme

This type corresponds to HTTP standard Related Schemes.

Traits

Host

HostFields

HostFieldsWithStore

HostRequest

HostRequestOptions

HostRequestOptionsWithStore

HostRequestWithStore

HostResponse

HostResponseWithStore

HostWithStore

Functions

add_to_linker

Type Aliases

Duration

FieldName

Field names are always strings.

FieldValue

Field values should always be ASCII strings. However, in reality, HTTP implementations often have to interpret malformed values, so they are provided as a list of bytes.

Headers

Headers is an alias for Fields.

StatusCode

This type corresponds to the HTTP standard Status Code.

Trailers

Trailers is an alias for Fields.