spin_world::wasi::http0_3_0_rc_2025_09_16::types

Trait HostRequest

pub trait HostRequest: Send {

    // Required methods
    fn new(
        &mut self,
        headers: Resource<Headers>,
        contents: Option<StreamReader<u8>>,
        trailers: FutureReader<Result<Option<Resource<Trailers>>, ErrorCode>>,
        options: Option<Resource<RequestOptions>>,
    ) -> impl Future<Output = Result<(Resource<Request>, FutureReader<Result<(), ErrorCode>>)>> + Send;
    fn get_method(
        &mut self,
        self_: Resource<Request>,
    ) -> impl Future<Output = Result<Method>> + Send;
    fn set_method(
        &mut self,
        self_: Resource<Request>,
        method: Method,
    ) -> impl Future<Output = Result<Result<(), ()>>> + Send;
    fn get_path_with_query(
        &mut self,
        self_: Resource<Request>,
    ) -> impl Future<Output = Result<Option<String>>> + Send;
    fn set_path_with_query(
        &mut self,
        self_: Resource<Request>,
        path_with_query: Option<String>,
    ) -> impl Future<Output = Result<Result<(), ()>>> + Send;
    fn get_scheme(
        &mut self,
        self_: Resource<Request>,
    ) -> impl Future<Output = Result<Option<Scheme>>> + Send;
    fn set_scheme(
        &mut self,
        self_: Resource<Request>,
        scheme: Option<Scheme>,
    ) -> impl Future<Output = Result<Result<(), ()>>> + Send;
    fn get_authority(
        &mut self,
        self_: Resource<Request>,
    ) -> impl Future<Output = Result<Option<String>>> + Send;
    fn set_authority(
        &mut self,
        self_: Resource<Request>,
        authority: Option<String>,
    ) -> impl Future<Output = Result<Result<(), ()>>> + Send;
    fn get_options(
        &mut self,
        self_: Resource<Request>,
    ) -> impl Future<Output = Result<Option<Resource<RequestOptions>>>> + Send;
    fn get_headers(
        &mut self,
        self_: Resource<Request>,
    ) -> impl Future<Output = Result<Resource<Headers>>> + Send;
    fn consume_body(
        &mut self,
        this: Resource<Request>,
        res: FutureReader<Result<(), ErrorCode>>,
    ) -> impl Future<Output = Result<(StreamReader<u8>, FutureReader<Result<Option<Resource<Trailers>>, ErrorCode>>)>> + Send;
    fn drop(
        &mut self,
        rep: Resource<Request>,
    ) -> impl Future<Output = Result<()>> + Send;

}

Required Methods

fn new( &mut self, headers: Resource<Headers>, contents: Option<StreamReader<u8>>, trailers: FutureReader<Result<Option<Resource<Trailers>>, ErrorCode>>, options: Option<Resource<RequestOptions>>, ) -> impl Future<Output = Result<(Resource<Request>, FutureReader<Result<(), ErrorCode>>)>> + Send

Construct a new request with a default method of GET, and none values for path-with-query, scheme, and authority.

headers is the HTTP Headers for the Request.

contents is the optional body content stream with none representing a zero-length content stream. Once it is closed, trailers future must resolve to a result. If trailers resolves to an error, underlying connection will be closed immediately.

options is optional request-options resource to be used if the request is sent over a network connection.

It is possible to construct, or manipulate with the accessor functions below, a request with an invalid combination of scheme and authority, or headers which are not permitted to be sent. It is the obligation of the handler.handle implementation to reject invalid constructions of request.

The returned future resolves to result of transmission of this request.

fn get_method( &mut self, self_: Resource<Request>, ) -> impl Future<Output = Result<Method>> + Send

Get the Method for the Request.

fn set_method( &mut self, self_: Resource<Request>, method: Method, ) -> impl Future<Output = Result<Result<(), ()>>> + Send

Set the Method for the Request. Fails if the string present in a method.other argument is not a syntactically valid method.

fn get_path_with_query( &mut self, self_: Resource<Request>, ) -> impl Future<Output = Result<Option<String>>> + Send

Get the combination of the HTTP Path and Query for the Request. When none, this represents an empty Path and empty Query.

fn set_path_with_query( &mut self, self_: Resource<Request>, path_with_query: Option<String>, ) -> impl Future<Output = Result<Result<(), ()>>> + Send

Set the combination of the HTTP Path and Query for the Request. When none, this represents an empty Path and empty Query. Fails is the string given is not a syntactically valid path and query uri component.

fn get_scheme( &mut self, self_: Resource<Request>, ) -> impl Future<Output = Result<Option<Scheme>>> + Send

Get the HTTP Related Scheme for the Request. When none, the implementation may choose an appropriate default scheme.

fn set_scheme( &mut self, self_: Resource<Request>, scheme: Option<Scheme>, ) -> impl Future<Output = Result<Result<(), ()>>> + Send

Set the HTTP Related Scheme for the Request. When none, the implementation may choose an appropriate default scheme. Fails if the string given is not a syntactically valid uri scheme.

fn get_authority( &mut self, self_: Resource<Request>, ) -> impl Future<Output = Result<Option<String>>> + Send

Get the authority of the Request’s target URI. A value of none may be used with Related Schemes which do not require an authority. The HTTP and HTTPS schemes always require an authority.

fn set_authority( &mut self, self_: Resource<Request>, authority: Option<String>, ) -> impl Future<Output = Result<Result<(), ()>>> + Send

Set the authority of the Request’s target URI. A value of none may be used with Related Schemes which do not require an authority. The HTTP and HTTPS schemes always require an authority. Fails if the string given is not a syntactically valid URI authority.

fn get_options( &mut self, self_: Resource<Request>, ) -> impl Future<Output = Result<Option<Resource<RequestOptions>>>> + Send

Get the request-options to be associated with this request

The returned request-options resource is immutable: set-* operations will fail if invoked.

This request-options resource is a child: it must be dropped before the parent request is dropped, or its ownership is transferred to another component by e.g. handler.handle.

fn get_headers( &mut self, self_: Resource<Request>, ) -> impl Future<Output = Result<Resource<Headers>>> + Send

Get the headers associated with the Request.

The returned headers resource is immutable: set, append, and delete operations will fail with header-error.immutable.

fn consume_body( &mut self, this: Resource<Request>, res: FutureReader<Result<(), ErrorCode>>, ) -> impl Future<Output = Result<(StreamReader<u8>, FutureReader<Result<Option<Resource<Trailers>>, ErrorCode>>)>> + Send

Get body of the Request.

Stream returned by this method represents the contents of the body. Once the stream is reported as closed, callers should await the returned future to determine whether the body was received successfully. The future will only resolve after the stream is reported as closed.

This function takes a res future as a parameter, which can be used to communicate an error in handling of the request.

Note that function will move the request, but references to headers or request options acquired from it previously will remain valid.

fn drop( &mut self, rep: Resource<Request>, ) -> impl Future<Output = Result<()>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.