spin_world::wasi::http0_3_0_rc_2025_09_16::types

Trait HostRequestOptions

pub trait HostRequestOptions: Send {
    // Required methods
    fn new(
        &mut self,
    ) -> impl Future<Output = Result<Resource<RequestOptions>>> + Send;
    fn get_connect_timeout(
        &mut self,
        self_: Resource<RequestOptions>,
    ) -> impl Future<Output = Result<Option<Duration>>> + Send;
    fn set_connect_timeout(
        &mut self,
        self_: Resource<RequestOptions>,
        duration: Option<Duration>,
    ) -> impl Future<Output = Result<Result<(), RequestOptionsError>>> + Send;
    fn get_first_byte_timeout(
        &mut self,
        self_: Resource<RequestOptions>,
    ) -> impl Future<Output = Result<Option<Duration>>> + Send;
    fn set_first_byte_timeout(
        &mut self,
        self_: Resource<RequestOptions>,
        duration: Option<Duration>,
    ) -> impl Future<Output = Result<Result<(), RequestOptionsError>>> + Send;
    fn get_between_bytes_timeout(
        &mut self,
        self_: Resource<RequestOptions>,
    ) -> impl Future<Output = Result<Option<Duration>>> + Send;
    fn set_between_bytes_timeout(
        &mut self,
        self_: Resource<RequestOptions>,
        duration: Option<Duration>,
    ) -> impl Future<Output = Result<Result<(), RequestOptionsError>>> + Send;
    fn clone(
        &mut self,
        self_: Resource<RequestOptions>,
    ) -> impl Future<Output = Result<Resource<RequestOptions>>> + Send;
    fn drop(
        &mut self,
        rep: Resource<RequestOptions>,
    ) -> impl Future<Output = Result<()>> + Send;
}

Required Methods

fn new( &mut self, ) -> impl Future<Output = Result<Resource<RequestOptions>>> + Send

Construct a default request-options value.

fn get_connect_timeout( &mut self, self_: Resource<RequestOptions>, ) -> impl Future<Output = Result<Option<Duration>>> + Send

The timeout for the initial connect to the HTTP Server.

fn set_connect_timeout( &mut self, self_: Resource<RequestOptions>, duration: Option<Duration>, ) -> impl Future<Output = Result<Result<(), RequestOptionsError>>> + Send

Set the timeout for the initial connect to the HTTP Server. An error return value indicates that this timeout is not supported or that this handle is immutable.

fn get_first_byte_timeout( &mut self, self_: Resource<RequestOptions>, ) -> impl Future<Output = Result<Option<Duration>>> + Send

The timeout for receiving the first byte of the Response body.

fn set_first_byte_timeout( &mut self, self_: Resource<RequestOptions>, duration: Option<Duration>, ) -> impl Future<Output = Result<Result<(), RequestOptionsError>>> + Send

Set the timeout for receiving the first byte of the Response body. An error return value indicates that this timeout is not supported or that this handle is immutable.

fn get_between_bytes_timeout( &mut self, self_: Resource<RequestOptions>, ) -> impl Future<Output = Result<Option<Duration>>> + Send

The timeout for receiving subsequent chunks of bytes in the Response body stream.

fn set_between_bytes_timeout( &mut self, self_: Resource<RequestOptions>, duration: Option<Duration>, ) -> impl Future<Output = Result<Result<(), RequestOptionsError>>> + Send

Set the timeout for receiving subsequent chunks of bytes in the Response body stream. An error return value indicates that this timeout is not supported or that this handle is immutable.

fn clone( &mut self, self_: Resource<RequestOptions>, ) -> impl Future<Output = Result<Resource<RequestOptions>>> + Send

Make a deep copy of the request-options. The resulting request-options is mutable.

fn drop( &mut self, rep: Resource<RequestOptions>, ) -> impl Future<Output = Result<()>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.