spin_world::wasi::io0_2_0::streams
Trait HostInputStream

pub trait HostInputStream {
    // Required methods
    fn read<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<InputStream>,
        len: u64,
    ) -> Pin<Box<dyn Future<Output = Result<Result<Vec<u8>, StreamError>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocking_read<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<InputStream>,
        len: u64,
    ) -> Pin<Box<dyn Future<Output = Result<Result<Vec<u8>, StreamError>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn skip<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<InputStream>,
        len: u64,
    ) -> Pin<Box<dyn Future<Output = Result<Result<u64, StreamError>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blocking_skip<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<InputStream>,
        len: u64,
    ) -> Pin<Box<dyn Future<Output = Result<Result<u64, StreamError>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<InputStream>,
    ) -> Pin<Box<dyn Future<Output = Result<Resource<Pollable>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn drop<'life0, 'async_trait>(
        &'life0 mut self,
        rep: Resource<InputStream>,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn read<'life0, 'async_trait>( &'life0 mut self, self_: Resource<InputStream>, len: u64, ) -> Pin<Box<dyn Future<Output = Result<Result<Vec<u8>, StreamError>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Perform a non-blocking read from the stream.

When the source of a read is binary data, the bytes from the source are returned verbatim. When the source of a read is known to the implementation to be text, bytes containing the UTF-8 encoding of the text are returned.

This function returns a list of bytes containing the read data, when successful. The returned list will contain up to len bytes; it may return fewer than requested, but not more. The list is empty when no bytes are available for reading at this time. The pollable given by subscribe will be ready when more bytes are available.

This function fails with a stream-error when the operation encounters an error, giving last-operation-failed, or when the stream is closed, giving closed.

When the caller gives a len of 0, it represents a request to read 0 bytes. If the stream is still open, this call should succeed and return an empty list, or otherwise fail with closed.

The len parameter is a u64, which could represent a list of u8 which is not possible to allocate in wasm32, or not desirable to allocate as as a return value by the callee. The callee may return a list of bytes less than len in size while more bytes are available for reading.

fn blocking_read<'life0, 'async_trait>( &'life0 mut self, self_: Resource<InputStream>, len: u64, ) -> Pin<Box<dyn Future<Output = Result<Result<Vec<u8>, StreamError>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Read bytes from a stream, after blocking until at least one byte can be read. Except for blocking, behavior is identical to read.

fn skip<'life0, 'async_trait>( &'life0 mut self, self_: Resource<InputStream>, len: u64, ) -> Pin<Box<dyn Future<Output = Result<Result<u64, StreamError>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Skip bytes from a stream. Returns number of bytes skipped.

Behaves identical to read, except instead of returning a list of bytes, returns the number of bytes consumed from the stream.

fn blocking_skip<'life0, 'async_trait>( &'life0 mut self, self_: Resource<InputStream>, len: u64, ) -> Pin<Box<dyn Future<Output = Result<Result<u64, StreamError>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Skip bytes from a stream, after blocking until at least one byte can be skipped. Except for blocking behavior, identical to skip.

fn subscribe<'life0, 'async_trait>( &'life0 mut self, self_: Resource<InputStream>, ) -> Pin<Box<dyn Future<Output = Result<Resource<Pollable>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Create a pollable which will resolve once either the specified stream has bytes available to read or the other end of the stream has been closed. The created pollable is a child resource of the input-stream. Implementations may trap if the input-stream is dropped before all derived pollables created with this function are dropped.

fn drop<'life0, 'async_trait>( &'life0 mut self, rep: Resource<InputStream>, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,