spin_world::wasi::keyvalue::atomics
Trait Host

pub trait Host: Send + HostCas {
    // Required methods
    fn increment<'life0, 'async_trait>(
        &'life0 mut self,
        bucket: Resource<Bucket>,
        key: String,
        delta: i64,
    ) -> Pin<Box<dyn Future<Output = Result<i64, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn swap<'life0, 'async_trait>(
        &'life0 mut self,
        cas: Resource<Cas>,
        value: Vec<u8>,
    ) -> Pin<Box<dyn Future<Output = Result<Result<(), CasError>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn increment<'life0, 'async_trait>( &'life0 mut self, bucket: Resource<Bucket>, key: String, delta: i64, ) -> Pin<Box<dyn Future<Output = Result<i64, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Atomically increment the value associated with the key in the store by the given delta. It returns the new value.

If the key does not exist in the store, it creates a new key-value pair with the value set to the given delta.

If any other error occurs, it returns an Err(error).

fn swap<'life0, 'async_trait>( &'life0 mut self, cas: Resource<Cas>, value: Vec<u8>, ) -> Pin<Box<dyn Future<Output = Result<Result<(), CasError>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Perform the swap on a CAS operation. This consumes the CAS handle and returns an error if the CAS operation failed.