spin_world::wasi::keyvalue::store

Trait HostBucket

pub trait HostBucket: Send + Send {
    // Required methods
    fn get(
        &mut self,
        self_: Resource<Bucket>,
        key: String,
    ) -> impl Future<Output = Result<Option<Vec<u8>>, Error>> + Send;
    fn set(
        &mut self,
        self_: Resource<Bucket>,
        key: String,
        value: Vec<u8>,
    ) -> impl Future<Output = Result<(), Error>> + Send;
    fn delete(
        &mut self,
        self_: Resource<Bucket>,
        key: String,
    ) -> impl Future<Output = Result<(), Error>> + Send;
    fn exists(
        &mut self,
        self_: Resource<Bucket>,
        key: String,
    ) -> impl Future<Output = Result<bool, Error>> + Send;
    fn list_keys(
        &mut self,
        self_: Resource<Bucket>,
        cursor: Option<String>,
    ) -> impl Future<Output = Result<KeyResponse, Error>> + Send;
    fn drop(
        &mut self,
        rep: Resource<Bucket>,
    ) -> impl Future<Output = Result<()>> + Send;
}

Required Methods

fn get( &mut self, self_: Resource<Bucket>, key: String, ) -> impl Future<Output = Result<Option<Vec<u8>>, Error>> + Send

Get the value associated with the specified key

The value is returned as an option. If the key-value pair exists in the store, it returns Ok(value). If the key does not exist in the store, it returns Ok(none).

If any other error occurs, it returns an Err(error).

fn set( &mut self, self_: Resource<Bucket>, key: String, value: Vec<u8>, ) -> impl Future<Output = Result<(), Error>> + Send

Set the value associated with the key in the store. If the key already exists in the store, it overwrites the value.

If the key does not exist in the store, it creates a new key-value pair.

If any other error occurs, it returns an Err(error).

fn delete( &mut self, self_: Resource<Bucket>, key: String, ) -> impl Future<Output = Result<(), Error>> + Send

Delete the key-value pair associated with the key in the store.

If the key does not exist in the store, it does nothing.

If any other error occurs, it returns an Err(error).

fn exists( &mut self, self_: Resource<Bucket>, key: String, ) -> impl Future<Output = Result<bool, Error>> + Send

Check if the key exists in the store.

If the key exists in the store, it returns Ok(true). If the key does not exist in the store, it returns Ok(false).

If any other error occurs, it returns an Err(error).

fn list_keys( &mut self, self_: Resource<Bucket>, cursor: Option<String>, ) -> impl Future<Output = Result<KeyResponse, Error>> + Send

Get all the keys in the store with an optional cursor (for use in pagination). It returns a list of keys. Please note that for most KeyValue implementations, this is a can be a very expensive operation and so it should be used judiciously. Implementations can return any number of keys in a single response, but they should never attempt to send more data than is reasonable (i.e. on a small edge device, this may only be a few KB, while on a large machine this could be several MB). Any response should also return a cursor that can be used to fetch the next page of keys. See the key-response record for more information.

Note that the keys are not guaranteed to be returned in any particular order.

If the store is empty, it returns an empty list.

MAY show an out-of-date list of keys if there are concurrent writes to the store.

If any error occurs, it returns an Err(error).

fn drop( &mut self, rep: Resource<Bucket>, ) -> impl Future<Output = Result<()>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.