spin_world::wasi::otel

Module metrics

Structs

Exemplar

A measurement sampled from a time series providing a typical example.

ExponentialBucket

A set of bucket counts, encoded in a contiguous array of counts.

ExponentialHistogram

The histogram of all measurements of values from an instrument.

ExponentialHistogramDataPoint

A single data point in a time series to be associated with an exponential-histogram .

Gauge

A measurement of the current value of an instrument.

GaugeDataPoint

A single data point in a time series to be associated with a gauge.

Histogram

Represents the histogram of all measurements of values from an instrument.

HistogramDataPoint

A single data point in a time series to be associated with a histogram.

Metric

A collection of one or more aggregated time series from a metric.

ResourceMetrics

A collection of scope-metrics and the associated resource that created them.

ScopeMetrics

A collection of metrics produced by a meter.

Sum

Represents the sum of all measurements of values from an instrument.

SumDataPoint

A single data point in a time series to be associated with a sum.

Enums

MetricData

Metric data for all types.

MetricNumber

The number types available for any given instrument.

Temporality

Defines the window that an aggregation was calculated over.

Traits

Host

HostWithStore

Functions

add_to_linker

Type Aliases

Datetime

Duration

Error

An error resulting from export being called.

InstrumentationScope

KeyValue

Resource

SpanId

TraceId