spin_world::wasi::sockets0_3_0_rc_2025_09_16::types

Trait HostTcpSocket

pub trait HostTcpSocket: Send {

    // Required methods
    fn create(
        &mut self,
        address_family: IpAddressFamily,
    ) -> impl Future<Output = Result<Result<Resource<TcpSocket>, ErrorCode>>> + Send;
    fn bind(
        &mut self,
        self_: Resource<TcpSocket>,
        local_address: IpSocketAddress,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn listen(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<Result<StreamReader<Resource<TcpSocket>>, ErrorCode>>> + Send;
    fn receive(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<(StreamReader<u8>, FutureReader<Result<(), ErrorCode>>)>> + Send;
    fn get_local_address(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<Result<IpSocketAddress, ErrorCode>>> + Send;
    fn get_remote_address(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<Result<IpSocketAddress, ErrorCode>>> + Send;
    fn get_is_listening(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<bool>> + Send;
    fn get_address_family(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<IpAddressFamily>> + Send;
    fn set_listen_backlog_size(
        &mut self,
        self_: Resource<TcpSocket>,
        value: u64,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn get_keep_alive_enabled(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<Result<bool, ErrorCode>>> + Send;
    fn set_keep_alive_enabled(
        &mut self,
        self_: Resource<TcpSocket>,
        value: bool,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn get_keep_alive_idle_time(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<Result<Duration, ErrorCode>>> + Send;
    fn set_keep_alive_idle_time(
        &mut self,
        self_: Resource<TcpSocket>,
        value: Duration,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn get_keep_alive_interval(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<Result<Duration, ErrorCode>>> + Send;
    fn set_keep_alive_interval(
        &mut self,
        self_: Resource<TcpSocket>,
        value: Duration,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn get_keep_alive_count(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<Result<u32, ErrorCode>>> + Send;
    fn set_keep_alive_count(
        &mut self,
        self_: Resource<TcpSocket>,
        value: u32,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn get_hop_limit(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<Result<u8, ErrorCode>>> + Send;
    fn set_hop_limit(
        &mut self,
        self_: Resource<TcpSocket>,
        value: u8,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn get_receive_buffer_size(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<Result<u64, ErrorCode>>> + Send;
    fn set_receive_buffer_size(
        &mut self,
        self_: Resource<TcpSocket>,
        value: u64,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn get_send_buffer_size(
        &mut self,
        self_: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<Result<u64, ErrorCode>>> + Send;
    fn set_send_buffer_size(
        &mut self,
        self_: Resource<TcpSocket>,
        value: u64,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn drop(
        &mut self,
        rep: Resource<TcpSocket>,
    ) -> impl Future<Output = Result<()>> + Send;

}

Required Methods

fn create( &mut self, address_family: IpAddressFamily, ) -> impl Future<Output = Result<Result<Resource<TcpSocket>, ErrorCode>>> + Send

Create a new TCP socket.

Similar to socket(AF_INET or AF_INET6, SOCK_STREAM, IPPROTO_TCP) in POSIX. On IPv6 sockets, IPV6_V6ONLY is enabled by default and can’t be configured otherwise.

Unlike POSIX, WASI sockets have no notion of a socket-level O_NONBLOCK flag. Instead they fully rely on the Component Model’s async support.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/socket.html

https://man7.org/linux/man-pages/man2/socket.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock2/nf-winsock2-wsasocketw

https://man.freebsd.org/cgi/man.cgi?query=socket&sektion=2

fn bind( &mut self, self_: Resource<TcpSocket>, local_address: IpSocketAddress, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Bind the socket to the provided IP address and port.

If the IP address is zero (0.0.0.0 in IPv4, :: in IPv6), it is left to the implementation to decide which network interface(s) to bind to. If the TCP/UDP port is zero, the socket will be bound to a random free port.

Bind can be attempted multiple times on the same socket, even with different arguments on each iteration. But never concurrently and only as long as the previous bind failed. Once a bind succeeds, the binding can’t be changed anymore.

Typical errors

invalid-argument: The local-address has the wrong address family. (EAFNOSUPPORT, EFAULT on Windows)

invalid-argument: local-address is not a unicast address. (EINVAL)

invalid-argument: local-address is an IPv4-mapped IPv6 address. (EINVAL)

invalid-state: The socket is already bound. (EINVAL)

address-in-use: No ephemeral ports available. (EADDRINUSE, ENOBUFS on Windows)

address-in-use: Address is already in use. (EADDRINUSE)

address-not-bindable: local-address is not an address that can be bound to. (EADDRNOTAVAIL)

Implementors note

When binding to a non-zero port, this bind operation shouldn’t be affected by the TIME_WAIT state of a recently closed socket on the same local address. In practice this means that the SO_REUSEADDR socket option should be set implicitly on all platforms, except on Windows where this is the default behavior and SO_REUSEADDR performs something different entirely.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/bind.html

https://man7.org/linux/man-pages/man2/bind.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock/nf-winsock-bind

https://man.freebsd.org/cgi/man.cgi?query=bind&sektion=2&format=html

fn listen( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<Result<StreamReader<Resource<TcpSocket>>, ErrorCode>>> + Send

Start listening and return a stream of new inbound connections.

Transitions the socket into the listening state. This can be called at most once per socket.

If the socket is not already explicitly bound, this function will implicitly bind the socket to a random free port.

Normally, the returned sockets are bound, in the connected state and immediately ready for I/O. Though, depending on exact timing and circumstances, a newly accepted connection may already be closed by the time the server attempts to perform its first I/O on it. This is true regardless of whether the WASI implementation uses “synthesized” sockets or not (see Implementors Notes below).

The following properties are inherited from the listener socket:

address-family

keep-alive-enabled

keep-alive-idle-time

keep-alive-interval

keep-alive-count

hop-limit

receive-buffer-size

send-buffer-size

Typical errors

invalid-state: The socket is already in the connected state. (EISCONN, EINVAL on BSD)

invalid-state: The socket is already in the listening state.

address-in-use: Tried to perform an implicit bind, but there were no ephemeral ports available. (EADDRINUSE)

Implementors note

This method returns a single perpetual stream that should only close on fatal errors (if any). Yet, the POSIX’ accept function may also return transient errors (e.g. ECONNABORTED). The exact details differ per operation system. For example, the Linux manual mentions:

Linux accept() passes already-pending network errors on the new socket as an error code from accept(). This behavior differs from other BSD socket implementations. For reliable operation the application should detect the network errors defined for the protocol after accept() and treat them like EAGAIN by retrying. In the case of TCP/IP, these are ENETDOWN, EPROTO, ENOPROTOOPT, EHOSTDOWN, ENONET, EHOSTUNREACH, EOPNOTSUPP, and ENETUNREACH. Source: https://man7.org/linux/man-pages/man2/accept.2.html

WASI implementations have two options to handle this:

Optionally log it and then skip over non-fatal errors returned by accept. Guest code never gets to see these failures. Or:

Synthesize a tcp-socket resource that exposes the error when attempting to send or receive on it. Guest code then sees these failures as regular I/O errors.

In either case, the stream returned by this listen method remains operational.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/listen.html

https://pubs.opengroup.org/onlinepubs/9699919799/functions/accept.html

https://man7.org/linux/man-pages/man2/listen.2.html

https://man7.org/linux/man-pages/man2/accept.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock2/nf-winsock2-listen

https://learn.microsoft.com/en-us/windows/win32/api/winsock2/nf-winsock2-accept

https://man.freebsd.org/cgi/man.cgi?query=listen&sektion=2

https://man.freebsd.org/cgi/man.cgi?query=accept&sektion=2

fn receive( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<(StreamReader<u8>, FutureReader<Result<(), ErrorCode>>)>> + Send

Read data from peer.

This function returns a stream which provides the data received from the socket, and a future providing additional error information in case the socket is closed abnormally.

If the socket is closed normally, stream.read on the stream will return read-status::closed with no error-context and the future resolves to the value ok. If the socket is closed abnormally, stream.read on the stream returns read-status::closed with an error-context and the future resolves to err with an error-code.

receive is meant to be called only once per socket. If it is called more than once, the subsequent calls return a new stream that fails as if it were closed abnormally.

If the caller is not expecting to receive any data from the peer, they may drop the stream. Any data still in the receive queue will be discarded. This is equivalent to calling shutdown(SHUT_RD) in POSIX.

Typical errors

invalid-state: The socket is not in the connected state. (ENOTCONN)

connection-reset: The connection was reset. (ECONNRESET)

remote-unreachable: The remote address is not reachable. (EHOSTUNREACH, EHOSTDOWN, ENETUNREACH, ENETDOWN, ENONET)

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/recv.html

https://man7.org/linux/man-pages/man2/recv.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock/nf-winsock-recv

https://man.freebsd.org/cgi/man.cgi?query=recv&sektion=2

fn get_local_address( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<Result<IpSocketAddress, ErrorCode>>> + Send

Get the bound local address.

POSIX mentions:

If the socket has not been bound to a local name, the value stored in the object pointed to by address is unspecified.

WASI is stricter and requires get-local-address to return invalid-state when the socket hasn’t been bound yet.

Typical errors

invalid-state: The socket is not bound to any local address.

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/getsockname.html

https://man7.org/linux/man-pages/man2/getsockname.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock/nf-winsock-getsockname

https://man.freebsd.org/cgi/man.cgi?getsockname

fn get_remote_address( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<Result<IpSocketAddress, ErrorCode>>> + Send

Get the remote address.

Typical errors

invalid-state: The socket is not connected to a remote address. (ENOTCONN)

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/getpeername.html

https://man7.org/linux/man-pages/man2/getpeername.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock/nf-winsock-getpeername

https://man.freebsd.org/cgi/man.cgi?query=getpeername&sektion=2&n=1

fn get_is_listening( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<bool>> + Send

Whether the socket is in the listening state.

Equivalent to the SO_ACCEPTCONN socket option.

fn get_address_family( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<IpAddressFamily>> + Send

Whether this is a IPv4 or IPv6 socket.

This is the value passed to the constructor.

Equivalent to the SO_DOMAIN socket option.

fn set_listen_backlog_size( &mut self, self_: Resource<TcpSocket>, value: u64, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Hints the desired listen queue size. Implementations are free to ignore this.

If the provided value is 0, an invalid-argument error is returned. Any other value will never cause an error, but it might be silently clamped and/or rounded.

Typical errors

not-supported: (set) The platform does not support changing the backlog size after the initial listen.

invalid-argument: (set) The provided value was 0.

invalid-state: (set) The socket is in the connecting or connected state.

fn get_keep_alive_enabled( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<Result<bool, ErrorCode>>> + Send

Enables or disables keepalive.

The keepalive behavior can be adjusted using:

keep-alive-idle-time

keep-alive-interval

keep-alive-count These properties can be configured while keep-alive-enabled is false, but only come into effect when keep-alive-enabled is true.

Equivalent to the SO_KEEPALIVE socket option.

fn set_keep_alive_enabled( &mut self, self_: Resource<TcpSocket>, value: bool, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

fn get_keep_alive_idle_time( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<Result<Duration, ErrorCode>>> + Send

Amount of time the connection has to be idle before TCP starts sending keepalive packets.

If the provided value is 0, an invalid-argument error is returned. Any other value will never cause an error, but it might be silently clamped and/or rounded. I.e. after setting a value, reading the same setting back may return a different value.

Equivalent to the TCP_KEEPIDLE socket option. (TCP_KEEPALIVE on MacOS)

Typical errors

invalid-argument: (set) The provided value was 0.

fn set_keep_alive_idle_time( &mut self, self_: Resource<TcpSocket>, value: Duration, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

fn get_keep_alive_interval( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<Result<Duration, ErrorCode>>> + Send

The time between keepalive packets.

If the provided value is 0, an invalid-argument error is returned. Any other value will never cause an error, but it might be silently clamped and/or rounded. I.e. after setting a value, reading the same setting back may return a different value.

Equivalent to the TCP_KEEPINTVL socket option.

Typical errors

invalid-argument: (set) The provided value was 0.

fn set_keep_alive_interval( &mut self, self_: Resource<TcpSocket>, value: Duration, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

fn get_keep_alive_count( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<Result<u32, ErrorCode>>> + Send

The maximum amount of keepalive packets TCP should send before aborting the connection.

If the provided value is 0, an invalid-argument error is returned. Any other value will never cause an error, but it might be silently clamped and/or rounded. I.e. after setting a value, reading the same setting back may return a different value.

Equivalent to the TCP_KEEPCNT socket option.

Typical errors

invalid-argument: (set) The provided value was 0.

fn set_keep_alive_count( &mut self, self_: Resource<TcpSocket>, value: u32, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

fn get_hop_limit( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<Result<u8, ErrorCode>>> + Send

Equivalent to the IP_TTL & IPV6_UNICAST_HOPS socket options.

If the provided value is 0, an invalid-argument error is returned.

Typical errors

invalid-argument: (set) The TTL value must be 1 or higher.

fn set_hop_limit( &mut self, self_: Resource<TcpSocket>, value: u8, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

fn get_receive_buffer_size( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<Result<u64, ErrorCode>>> + Send

The kernel buffer space reserved for sends/receives on this socket.

If the provided value is 0, an invalid-argument error is returned. Any other value will never cause an error, but it might be silently clamped and/or rounded. I.e. after setting a value, reading the same setting back may return a different value.

Equivalent to the SO_RCVBUF and SO_SNDBUF socket options.

Typical errors

invalid-argument: (set) The provided value was 0.

fn set_receive_buffer_size( &mut self, self_: Resource<TcpSocket>, value: u64, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

fn get_send_buffer_size( &mut self, self_: Resource<TcpSocket>, ) -> impl Future<Output = Result<Result<u64, ErrorCode>>> + Send

fn set_send_buffer_size( &mut self, self_: Resource<TcpSocket>, value: u64, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

fn drop( &mut self, rep: Resource<TcpSocket>, ) -> impl Future<Output = Result<()>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.