spin_world::wasi::sockets0_3_0_rc_2025_09_16::types

Trait HostUdpSocketWithStore

pub trait HostUdpSocketWithStore: HasData + Send {
    // Required methods
    fn send<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<UdpSocket>,
        data: Vec<u8>,
        remote_address: Option<IpSocketAddress>,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn receive<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<UdpSocket>,
    ) -> impl Future<Output = Result<Result<(Vec<u8>, IpSocketAddress), ErrorCode>>> + Send;
}

Required Methods

fn send<T>( accessor: &Accessor<T, Self>, self_: Resource<UdpSocket>, data: Vec<u8>, remote_address: Option<IpSocketAddress>, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Send a message on the socket to a particular peer.

If the socket is connected, the peer address may be left empty. In that case this is equivalent to send in POSIX. Otherwise it is equivalent to sendto.

Additionally, if the socket is connected, a remote-address argument may be provided but then it must be identical to the address passed to connect.

Implementations may trap if the data length exceeds 64 KiB.

Typical errors

invalid-argument: The remote-address has the wrong address family. (EAFNOSUPPORT)

invalid-argument: The IP address in remote-address is set to INADDR_ANY (0.0.0.0 / ::). (EDESTADDRREQ, EADDRNOTAVAIL)

invalid-argument: The port in remote-address is set to 0. (EDESTADDRREQ, EADDRNOTAVAIL)

invalid-argument: The socket is in “connected” mode and remote-address is some value that does not match the address passed to connect. (EISCONN)

invalid-argument: The socket is not “connected” and no value for remote-address was provided. (EDESTADDRREQ)

remote-unreachable: The remote address is not reachable. (ECONNRESET, ENETRESET on Windows, EHOSTUNREACH, EHOSTDOWN, ENETUNREACH, ENETDOWN, ENONET)

connection-refused: The connection was refused. (ECONNREFUSED)

datagram-too-large: The datagram is too large. (EMSGSIZE)

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/sendto.html

https://pubs.opengroup.org/onlinepubs/9699919799/functions/sendmsg.html

https://man7.org/linux/man-pages/man2/send.2.html

https://man7.org/linux/man-pages/man2/sendmmsg.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock2/nf-winsock2-send

https://learn.microsoft.com/en-us/windows/win32/api/winsock2/nf-winsock2-sendto

https://learn.microsoft.com/en-us/windows/win32/api/winsock2/nf-winsock2-wsasendmsg

https://man.freebsd.org/cgi/man.cgi?query=send&sektion=2

fn receive<T>( accessor: &Accessor<T, Self>, self_: Resource<UdpSocket>, ) -> impl Future<Output = Result<Result<(Vec<u8>, IpSocketAddress), ErrorCode>>> + Send

Receive a message on the socket.

On success, the return value contains a tuple of the received data and the address of the sender. Theoretical maximum length of the data is 64 KiB. Though in practice, it will typically be less than 1500 bytes.

If the socket is connected, the sender address is guaranteed to match the remote address passed to connect.

Typical errors

invalid-state: The socket has not been bound yet.

remote-unreachable: The remote address is not reachable. (ECONNRESET, ENETRESET on Windows, EHOSTUNREACH, EHOSTDOWN, ENETUNREACH, ENETDOWN, ENONET)

connection-refused: The connection was refused. (ECONNREFUSED)

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/recvfrom.html

https://pubs.opengroup.org/onlinepubs/9699919799/functions/recvmsg.html

https://man7.org/linux/man-pages/man2/recv.2.html

https://man7.org/linux/man-pages/man2/recvmmsg.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock/nf-winsock-recvfrom

https://learn.microsoft.com/en-us/windows/win32/api/mswsock/nc-mswsock-lpfn_wsarecvmsg

https://man.freebsd.org/cgi/man.cgi?query=recv&sektion=2

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.