spin_world::wasi::sockets0_3_0_rc_2026_01_06::types

Trait HostTcpSocketWithStore

pub trait HostTcpSocketWithStore: HasData + Send {
    // Required methods
    fn connect<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<TcpSocket>,
        remote_address: IpSocketAddress,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
    fn send<T>(
        accessor: &Accessor<T, Self>,
        self_: Resource<TcpSocket>,
        data: StreamReader<u8>,
    ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send;
}

Required Methods

fn connect<T>( accessor: &Accessor<T, Self>, self_: Resource<TcpSocket>, remote_address: IpSocketAddress, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Connect to a remote endpoint.

On success, the socket is transitioned into the connected state and this function returns a connection resource.

After a failed connection attempt, the socket will be in the closed state and the only valid action left is to drop the socket. A single socket can not be used to connect more than once.

Typical errors

invalid-argument: The remote-address has the wrong address family. (EAFNOSUPPORT)

invalid-argument: remote-address is not a unicast address. (EINVAL, ENETUNREACH on Linux, EAFNOSUPPORT on MacOS)

invalid-argument: remote-address is an IPv4-mapped IPv6 address. (EINVAL, EADDRNOTAVAIL on Illumos)

invalid-argument: The IP address in remote-address is set to INADDR_ANY (0.0.0.0 / ::). (EADDRNOTAVAIL on Windows)

invalid-argument: The port in remote-address is set to 0. (EADDRNOTAVAIL on Windows)

invalid-state: The socket is already in the connecting state. (EALREADY)

invalid-state: The socket is already in the connected state. (EISCONN)

invalid-state: The socket is already in the listening state. (EOPNOTSUPP, EINVAL on Windows)

timeout: Connection timed out. (ETIMEDOUT)

connection-refused: The connection was forcefully rejected. (ECONNREFUSED)

connection-reset: The connection was reset. (ECONNRESET)

connection-aborted: The connection was aborted. (ECONNABORTED)

remote-unreachable: The remote address is not reachable. (EHOSTUNREACH, EHOSTDOWN, ENETUNREACH, ENETDOWN, ENONET)

address-in-use: Tried to perform an implicit bind, but there were no ephemeral ports available. (EADDRINUSE, EADDRNOTAVAIL on Linux, EAGAIN on BSD)

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/connect.html

https://man7.org/linux/man-pages/man2/connect.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock2/nf-winsock2-connect

https://man.freebsd.org/cgi/man.cgi?connect

fn send<T>( accessor: &Accessor<T, Self>, self_: Resource<TcpSocket>, data: StreamReader<u8>, ) -> impl Future<Output = Result<Result<(), ErrorCode>>> + Send

Transmit data to peer.

The caller should close the stream when it has no more data to send to the peer. Under normal circumstances this will cause a FIN packet to be sent out. Closing the stream is equivalent to calling shutdown(SHUT_WR) in POSIX.

This function may be called at most once and returns once the full contents of the stream are transmitted or an error is encountered.

Typical errors

invalid-state: The socket is not in the connected state. (ENOTCONN)

connection-reset: The connection was reset. (ECONNRESET)

remote-unreachable: The remote address is not reachable. (EHOSTUNREACH, EHOSTDOWN, ENETUNREACH, ENETDOWN, ENONET)

References

https://pubs.opengroup.org/onlinepubs/9699919799/functions/send.html

https://man7.org/linux/man-pages/man2/send.2.html

https://learn.microsoft.com/en-us/windows/win32/api/winsock2/nf-winsock2-send

https://man.freebsd.org/cgi/man.cgi?query=send&sektion=2

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.