Crate spin_world

Re-exports

pub use fermyon::spin as v1;

pub use fermyon::spin2_0_0 as v2;

Modules

exports

fermyon

spin

wasi

Structs

Host_

Auto-generated bindings for an instance a component which implements the world host.

Host_Indices

Auto-generated bindings for index of the exports of host.

Host_Pre

Auto-generated bindings for a pre-instantiated version of a component which implements the world host.

LinkOptions

Link-time configurations.

Attribute Macros

async_trait