spin_world::fermyon::spin::sqlite

Trait Host

pub trait Host: Send {
    // Required methods
    fn open(
        &mut self,
        database: String,
    ) -> impl Future<Output = Result<Connection, Error>> + Send;
    fn execute(
        &mut self,
        conn: Connection,
        statement: String,
        parameters: Vec<Value>,
    ) -> impl Future<Output = Result<QueryResult, Error>> + Send;
    fn close(
        &mut self,
        conn: Connection,
    ) -> impl Future<Output = Result<()>> + Send;
    fn convert_error(&mut self, err: Error) -> Result<Error>;
}

Required Methods

fn open( &mut self, database: String, ) -> impl Future<Output = Result<Connection, Error>> + Send

Open a connection to a named database instance.

If database is “default”, the default instance is opened.

error::no-such-database will be raised if the name is not recognized.

fn execute( &mut self, conn: Connection, statement: String, parameters: Vec<Value>, ) -> impl Future<Output = Result<QueryResult, Error>> + Send

Execute a statement returning back data if there is any

fn close(&mut self, conn: Connection) -> impl Future<Output = Result<()>> + Send

Close the specified connection.

fn convert_error(&mut self, err: Error) -> Result<Error>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.