spin_world::wasi::http0_2_0::types

Trait HostIncomingRequest

pub trait HostIncomingRequest: Send {
    // Required methods
    fn method(
        &mut self,
        self_: Resource<IncomingRequest>,
    ) -> impl Future<Output = Result<Method>> + Send;
    fn path_with_query(
        &mut self,
        self_: Resource<IncomingRequest>,
    ) -> impl Future<Output = Result<Option<String>>> + Send;
    fn scheme(
        &mut self,
        self_: Resource<IncomingRequest>,
    ) -> impl Future<Output = Result<Option<Scheme>>> + Send;
    fn authority(
        &mut self,
        self_: Resource<IncomingRequest>,
    ) -> impl Future<Output = Result<Option<String>>> + Send;
    fn headers(
        &mut self,
        self_: Resource<IncomingRequest>,
    ) -> impl Future<Output = Result<Resource<Headers>>> + Send;
    fn consume(
        &mut self,
        self_: Resource<IncomingRequest>,
    ) -> impl Future<Output = Result<Result<Resource<IncomingBody>, ()>>> + Send;
    fn drop(
        &mut self,
        rep: Resource<IncomingRequest>,
    ) -> impl Future<Output = Result<()>> + Send;
}

Required Methods

fn method( &mut self, self_: Resource<IncomingRequest>, ) -> impl Future<Output = Result<Method>> + Send

Returns the method of the incoming request.

fn path_with_query( &mut self, self_: Resource<IncomingRequest>, ) -> impl Future<Output = Result<Option<String>>> + Send

Returns the path with query parameters from the request, as a string.

fn scheme( &mut self, self_: Resource<IncomingRequest>, ) -> impl Future<Output = Result<Option<Scheme>>> + Send

Returns the protocol scheme from the request.

fn authority( &mut self, self_: Resource<IncomingRequest>, ) -> impl Future<Output = Result<Option<String>>> + Send

Returns the authority from the request, if it was present.

fn headers( &mut self, self_: Resource<IncomingRequest>, ) -> impl Future<Output = Result<Resource<Headers>>> + Send

Get the headers associated with the request.

The returned headers resource is immutable: set, append, and delete operations will fail with header-error.immutable.

The headers returned are a child resource: it must be dropped before the parent incoming-request is dropped. Dropping this incoming-request before all children are dropped will trap.

fn consume( &mut self, self_: Resource<IncomingRequest>, ) -> impl Future<Output = Result<Result<Resource<IncomingBody>, ()>>> + Send

Gives the incoming-body associated with this request. Will only return success at most once, and subsequent calls will return error.

fn drop( &mut self, rep: Resource<IncomingRequest>, ) -> impl Future<Output = Result<()>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.