spin_world::wasi::keyvalue::atomics

Trait Host

pub trait Host: HostCas + Send {
    // Required methods
    fn increment(
        &mut self,
        bucket: Resource<Bucket>,
        key: String,
        delta: i64,
    ) -> impl Future<Output = Result<i64, Error>> + Send;
    fn swap(
        &mut self,
        cas: Resource<Cas>,
        value: Vec<u8>,
    ) -> impl Future<Output = Result<(), CasError>> + Send;
    fn convert_cas_error(&mut self, err: CasError) -> Result<CasError>;
}

Required Methods

fn increment( &mut self, bucket: Resource<Bucket>, key: String, delta: i64, ) -> impl Future<Output = Result<i64, Error>> + Send

Atomically increment the value associated with the key in the store by the given delta. It returns the new value.

If the key does not exist in the store, it creates a new key-value pair with the value set to the given delta.

If any other error occurs, it returns an Err(error).

fn swap( &mut self, cas: Resource<Cas>, value: Vec<u8>, ) -> impl Future<Output = Result<(), CasError>> + Send

Perform the swap on a CAS operation. This consumes the CAS handle and returns an error if the CAS operation failed.

fn convert_cas_error(&mut self, err: CasError) -> Result<CasError>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.